requirements.txt
requests
beautifulsoup4
gspread
google-auth
internetarchive
yt-dlp
